requirements.txt
langchain>=0.1.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=10.0.0
opencv-python>=4.8.0
pytube>=15.0.0
neo4j>=5.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
llama-cpp-python>=0.2.0